### 2018-05-11 #11 @whedon
[PRE REVIEW]: A short course about fitting models with the `scipy.optimize` module
**Submitting author:** @arokem (<a href="http://orcid.org/0000-0003-0679-1985">Ariel Rokem</a>)
**Repository:** <a href="https://github.com/arokem/scipy-optimize" target ="_blank">https://github.com/arokem/scipy-optimize</a>
**Version:** 0.1
