### 2018-05-25 #11 @arokem
From a quick read of their reviewing interests, I'd suggest Lindsey Heagy, Sidafa Conde, Thomas Arlidsen, and/or Manjari Narayan as reviewers.

Sadly, the google sheet of potential reviewers doesn't have their Github users, so I couldn't tag them here, see https://github.com/openjournals/jose/issues/12
