### 2018-05-25 #11 @labarba
Hang on, @arokem — we had a telecon with the editorial board last week, and have some more precise guidance on what we want in the paper. Stay tuned!
