### 2018-05-28 #11 @arokem
Will do.

Thanks for the guidance!

On Sat, May 26, 2018 at 3:22 PM, Lorena A. Barba <notifications@github.com>
wrote:

> @arokem <https://github.com/arokem> -- I modified our guidance notes a
> bit to reflect the discussion at our board telecon of a few days ago. See
> recent changes on the paper contents
> <https://github.com/openjournals/jose/commit/0e9d9fe16ee9272a2a89717ac4fb938861dbce65>
> and the reviewer checklist
> <https://github.com/openjournals/jose/commit/e9078a1f7abe9e9998f5b0cc06166c4d28b21bc7>
> .
>
> We'd like the JOSE paper to give more information to the readers, helping
> them quickly decide if they may want to investigate further by going to the
> repo, because they might be interested in adopting the materials. We should
> learn about the module at least:
>
>    - what need does it satisfy? how is it eligible for JOSE?
>    - how has it been used? how would someone else use it?
>    - what's the story behind it? what's the effort the author(s) have put
>    in and their expertise to teach this?
>
> Could you work some more on your paper, to this expanded guideline?
>
