+++ paper.bib
@@ -1,4 +1,4 @@
-@article {Rokem091553,
+@article {rokem091553,
 	author = {Rokem, Ariel and Landau, Ayelet Nina},
 	title = {The interaction of orientation-specific surround suppression and visual-spatial attention},
 	year = {2016},
@@ -6,4 +6,12 @@
 	URL = {https://www.biorxiv.org/content/early/2016/12/04/091553},
 	eprint = {https://www.biorxiv.org/content/early/2016/12/04/091553.full.pdf},
 	journal = {bioRxiv}
+}
+
+@Misc{scipy,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--},
+  url = "http://www.scipy.org/",
+  note = {[Online; accessed <today>]}
 }
