+++ paper.md
@@ -3,7 +3,7 @@ title: 'A short course about fitting models with the `scipy.optimize` module'
 tags:
 - scipy
 - optimization
-- models 
+- models
 authors:
 - name: Ariel Rokem
   orcid: 0000-0003-0679-1985
@@ -15,25 +15,97 @@ date: 23 April 2018
 bibliography: paper.bib
 ---
 
-# Summary 
+# Description
 
-Fitting models and testing the match of the models to the measured data is a fundamental 
-activity in many fields of science. This short (approximately 3-hour) course aims to teach 
-participants to use the Scipy library's  `optimize` module to fit models to data [@scipy]. 
-Using data from a psychology experiment [@rokem091553] as an example, the course motivates the 
-use of explicit mathematical models to  explain and predict data and compares linear models 
-and non-linear models. The core of the  lesson focuses on fitting a curve with the `curve_fit` 
-function. Finally, the course introduces the idea of model comparison with cross-validation
-for evaluation and selection between non-nested non-linear models. A video of an example of instructing this lesson by the author is available on [YouTube](https://www.youtube.com/watch?v=0eFokR-ikaA)
- 
-# Statement of need 
+Fitting models and testing the match of the models to the measured data
+is a fundamental activity in many fields of science. This short
+(approximately 3-hour) course (available at:
+[https://github.com/arokem/scipy-optimize](https://github.com/arokem/scipy-optimize))
+aims to teach participants to use the Scipy library's  `optimize` module
+to fit models to data [@scipy]. Using data from a psychology experiment
+[@rokem091553] as an example, the course motivates the use of explicit
+mathematical models to explain and predict data and compares linear
+models and non-linear models. The core of the lesson focuses on fitting a
+curve with the `curve_fit` function. The course also introduces the
+idea of model comparison with cross-validation for evaluation and
+selection between non-nested non-linear models.
 
-The target audience for this course are researchers or students with some programming 
-knowledge (e.g., having participated in a 'Software Carpentry' workshop beforehand). 
-Model fitting is useful in many different fields of research, but optimization for model
-fitting is not a topic that is usually covered in introductory statistics or computing 
-classes in. This course fills an existing need for hands-on curriculum that goes beyond the 
-topics taught in introductory computing workshops, such as 'Software Carpentry', providing 
-material for follow-up workshops on advanced/intermediate topics. 
+# Statement of need
+
+Model fitting is useful in many different fields of research, but
+optimization for model fitting is not a topic that is usually covered in
+introductory statistics or computing classes in many fields (e.g.,
+psychology). This course fills an existing need for hands-on curriculum
+that goes beyond the topics taught in introductory computing workshops,
+such as 'Software Carpentry', providing material for follow-up workshops
+on advanced/intermediate topics. The target audience for this course are
+researchers or students with some programming knowledge (e.g., having
+participated in a 'Software Carpentry' workshop beforehand).
+
+# Learning objectives
+
+In addition to these general objectives of this lessson, specific learning
+objectives are defined for each part of the lesson:
+
+Part 1:
+
+* Learners can define what a model is.
+* Learners can define model parameters and model fitting.
+* Learners can restate the benefits of modeling
+* Learners can explain the utility of modeling applied to their data.
+
+Part 2:
+
+* Learners can identify a linear model.
+* Learners can use `numpy` to fit a linear model to data
+* Learners can evaluate a model using model residuals.
+
+Part 3:
+
+* Learners can identify a nonlinear model, and discuss the differences
+  between linear and nonlinear models
+* Learners can use `scipy.optimize` to fit a nonlinear model to data.
[…]
