### 2018-05-29 #11 @arokem
I believe that changes implemented in https://github.com/arokem/scipy-optimize/commit/3f0810b74f5446f7a242fc793dcafa5b54514b90 should cover these requirements.
