### 2018-06-11 #11 @labarba
@arokem — I'm looking for reviewer(s) for your submission now. In the meantime, I have the following direction for you:

- Move "contributing" and "reporting issues" out of the paper and into a README in your repo.
- Let's re-organize the structure of the paper a little bit. For your fist paragraph, use the heading "Summary," instead of "Description." Then you have "Statement of Need" and "Learning Objectives," which are fine. Then, you could have a top heading "Description of the module," under which "Usage" can be a sub-heading, but put that after other sub-heading content (next item).
- The long heading "Content, instructional design, and experience of use in teaching and learning situations" is a little ugly; this comes straight from the Author Guide, I realize, but we didn't mean for it to be a heading, just direction on what to include. Yet, the content under this heading is scant. So, reorganize this under the heading "Description of the module," and tell us the story of this tutorial.

> This tutorial began as a blog post blah-blah Software Carptentry [link]. It was taught as a synchronous online workshop as part of the series blah-blah UC Davis [link]. The example data in the tutorial comes from an experiment in visual neuroscience [reference]. Learners follow along as we read the data from a comma-separated file, plot the data, then transform the data to increasingly reveal its structure. Then follows a lesson on different kinds of models blah-blah linear/nonlinear. We define the functional form for our model [reference] and a measure of modeling error. Blah-blah optimize. Blah evaluate. Poof, done.

Get my drift? Walk us through the tutorial in one brief paragraph, so we imagine what it is to sit through it.

If you have anything to add about _why_ you broke things down like you did, please tell us. This is the "instructional design" question. Is there a concept that learners find particularly difficult here, and you spend more time on it? And so on.
