### 2018-06-11 #11 @labarba
The [Author Guide](http://jose.theoj.org/about#author_guidelines) says:

> Computational learning modules should be complete and immediately usable for self-learning or adoption by other instructors.

Essentially, what we want the JOSE paper to do is to show that the submission meets the eligibility requirement. So it should explain how an instructor other than the author might adopt the module, or how an independent learner might use it.
