+++ README.md
@@ -1,0 +1,17 @@
+# scipy-optimize
+
+This is a short course about fitting models with the `scipy.optimize` module. For documentation see [paper](paper.md)
+
+## Contributing
+
+### Contributing to the module
+
+Contributions, corrections and improvements are welcome through a pull
+request against the repository at
+[https://github.com/arokem/scipy-optimize](https://github.com/arokem/scipy-optimize)
+
+### Reporting issues or problems with the module and seeking support
+
+Issues and problems with the module, as well as requests for support,
+should be submitted using the GitHub repo Issues page at:
+[https://github.com/arokem/scipy-optimize/issues](https://github.com/arokem/scipy-optimize/issues)
