### 2018-06-15 dbc794b
Edit paper with review suggestions.
