### 2018-06-15 #11 @arokem
Thanks for the comments @labarba!

Changes implemented in
https://github.com/arokem/scipy-optimize/commit/38c1d52b2fe7605f93e1a4dd45d09a05275b759b
and
https://github.com/arokem/scipy-optimize/commit/dbc794b8522261d29259100992b29495b3a73212
