### 2018-06-16 #16 @whedon
[REVIEW]: A short course about fitting models with the `scipy.optimize` module
**Submitting author:** @arokem (<a href="http://orcid.org/0000-0003-0679-1985">Ariel Rokem</a>)
**Repository:** <a href="https://github.com/arokem/scipy-optimize" target ="_blank">https://github.com/arokem/scipy-optimize</a>
**Version:** 0.1

- [x] **Version:** Does the release version given match the repository release (0.1)?
- [x] **Authorship:** Has the submitting author (@arokem) made substantial contributions to the module? Does the full list of authors seem appropriate and complete?

- [x] **Example usage:** Do the authors include examples of how to use the module?

### Pedagogy / Instructional design

- [x] **Version:** Does the release version given match the repository release (0.1)?
- [x] **Authorship:** Has the submitting author (@arokem) made substantial contributions to the module? Does the full list of authors seem appropriate and complete?

- [x] **Example usage:** Do the authors include examples of how to use the module?
