### 2018-06-16 #16 @labarba
👋 @oliviaguest, @ThomasA — Thank you for agreeing to review this submission to JOSE. Have a look at the [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines), and feel free to ask any questions here.

Each of you has a reviewer checklist at the top of this issue thread. You should check off each item, as you complete your review. You're also encouraged to open new issues on the submission's repository, as needed, adding a reference to them here.
