### 2018-06-23 arokem/scipy-optimize#4 @oliviaguest
Archival references should have a DOI
There are a few references to journal articles and it seems many (all?) are not referenced with their DOI but with a hyperlink to the PDF or to the journal website (which in the case of "work by Yu and Levi" in lesson 3 leads to a 404).
