### 2018-06-23 #16 @oliviaguest
Hi @arokem @labarba — I am not sure how to go about this, but if I understand correctly all items in my checklist are present/addressed except the final one. My issue is here: https://github.com/arokem/scipy-optimize/issues/4
