### 2018-06-27 4a9543e
Links/DOIs in reference section.
