+++ reference.md
@@ -1,13 +1,13 @@
 ---
 layout: page
-title: Lesson Title
+title: Reference
 subtitle: Reference
 ---
 
 
 Additional reading:
 
-Data analysis recipes: Fitting a model to data, Hogg et al.:  http://arxiv.org/abs/1008.4686
+[Data analysis recipes: Fitting a model to data, Hogg et al.](http://arxiv.org/abs/1008.4686)
 
 [An Introduction to Statistical Learning with Applications in R](http://www-bcf.usc.edu/~gareth/ISL/), by Gareth James,
 Daniela Witten, Trevor Hastie and Robert Tibshirani contains additional learning
@@ -18,7 +18,7 @@ The book is also accompanied by code written in R.
 Robert Tisbshirani and Jerome Friedman is a more heavily mathematical treatment
 of similar topics:
 
-Leo Breiman's [The two cultures of statistical modeling]( [https://projecteuclid.org/euclid.ss/1009213726) makes some observations about
+Leo Breiman's [The two cultures of statistical modeling](https://projecteuclid.org/euclid.ss/1009213726) makes some observations about
 different approaches to models, their predictive, and explanatory power.
 
 
