### 2018-06-27 arokem/scipy-optimize#4 @arokem
Good catch.

Addressed in https://github.com/arokem/scipy-optimize/commit/07341ccd4f73d941cfdb00dc86a03fc2ddcd1254 and https://github.com/arokem/scipy-optimize/commit/4a9543e214b3653d39f73c9415183e24b9777960
