### 2018-06-28 #16 @ThomasA
@labarba, is this a software submission or a learning module submission ([your reviewer guidelines](https://jose.theoj.org/about#reviewer_guidelines))? I assume a learning module?
