### 2018-06-28 #16 @ThomasA
If it's OK by you I am just going to comment here bit by bit as I go through the review.
