### 2018-06-28 #16 @ThomasA
- Authorship: @arokem has made substantial contributions to the repository and is the sole author listed on the paper. A number of other contributors have made relatively minor contributions; I guess it is OK that these are not authors?

  @gvwilson is also a substantial contributor to the repository - in fact by far the major contributor. At his point it is not clear to me which parts of the repository he contributed and whether he ought to be a co-author on the paper as well. Can @arokem and @gvwilson please clarify?
