### 2018-06-28 #16 @ThomasA
- Example usage: as I see it, we can inspect this and several other aspects on two levels:
  - Being a teaching module, one level is examples given to students of the module when taught.
  - The other level is examples of using the teaching module for other potential teachers.

  I assume the second level (meta-level) is the most important in this review procedure.

  As such, I do not think this submission provides much guidance in terms of examples of how to actually use the module. Here I am thinking that it might not for example be immediately obvious to new adopters of the module that the content seems to be set up to generated through GitHub's jekyll and how to use that. In particular, the file [instructors.md](https://github.com/arokem/scipy-optimize/blob/gh-pages/instructors.md) in the repository is empty - except for some headlines. Perhaps the module follows some common structure from Software Carpentry that has some generic usage documentation for instructors that could be referenced?
