### 2018-06-28 #16 @oliviaguest
@ThomasA have you seen the materials at the repo's URL? http://arokem.github.io/scipy-optimize/
