### 2018-06-28 #16 @ThomasA
@oliviaguest yes, I guess that can be seen as "getting started" advice for both learners and instructors alike, but I was thinking if there should be more "meta-instruction" on which practicalities are involved if you want to clone and modify the material for your own purposes? It could be a simple matter of making potential instructors aware that the material relies on GitHub jekyll for generation and referring to GitHub's documentation for that. Or maybe Software Carpentry has some generic instructions for their material  that applies here?
