### 2018-06-28 bdf9d92
Typo corrections for the linear models section
