### 2018-06-28 #16 @ThomasA
Apart from my above comment on usage examples, that I would like to discuss, I consider my review complete and can recommend publication.
