### 2018-06-28 arokem/scipy-optimize#4 @oliviaguest
I am just going to tag @labarba just in case there are any specific formatting instructions you and I should be aware of. :sweat_smile:

As far as I see this is addressed though as the hyperlinks all use a URL based on the DOI and so will resolve (much more easily) for readers.
