### 2018-06-28 #16 @labarba
I agree with @ThomasA. The Author Guidelines say:

> Computational learning modules should be complete and immediately usable for self-learning or adoption by other instructors.

In the paper (and possibly also the documentation), JOSE authors should explain how an instructor might adapt the module or how an independent learner might use it.

Readers also want to know _why_ they might adopt/use the module. This has to do with the approach to teaching a topic, and the learning scenario.

(Bear in mind, we're looking at the first few submissions to JOSE, and we are solidifying our genre as we hold these very conversations.)
