### 2018-07-02 873917a
recarray => Pandas DataFrame.
