### 2018-07-02 f6b2ae9
Add link to the website in the README. Add link to Carpentries template.
