### 2018-07-02 #16 @arokem
Indeed, I used the Carpentries (then Software Carpentry) lesson template to
set up the pages for this lesson.

I added a link to the Carpentries example lesson in the README in
https://github.com/arokem/scipy-optimize/commit/f6b2ae9dbded8e2610900292a439bf6e369c9a8a

On Thu, Jun 28, 2018 at 5:53 AM Lorena A. Barba <notifications@github.com>
wrote:

> I agree with @ThomasA <https://github.com/ThomasA>. The Author Guidelines
> say:
>
> Computational learning modules should be complete and immediately usable
> for self-learning or adoption by other instructors.
>
> In the paper (and possibly also the documentation), JOSE authors should
> explain how an instructor might adapt the module or how an independent
> learner might use it.
>
> Does that address this issue?

> Readers also want to know *why* they might adopt/use the module. This has
> to do with the approach to teaching a topic, and the learning scenario.
>
> I don't know what this means. Could you give an example (e.g., from
another paper) demonstrating this?

> (Bear in mind, we're looking at the first few submissions to JOSE, and we
> are solidifying our genre as we hold these very conversations.)
>
