### 2018-07-02 #16 @labarba
I see that only two check-boxes remain unticked in the review checklists.

Are we waiting on a couple of improvements from the author, at this point?
