### 2018-07-02 #16 @ThomasA
I think linking to the Software Carpentry lesson introduction has helped http://carpentries.github.io/lesson-example/. The submission is OK by me now.
