### 2018-07-03 #16 @labarba
@arokem Tiny fix: `SciPy` has a capital P (paper).
