### 2018-07-03 #16 @labarba
@arokem More fixes:

* Summary: In the reference to Jones at al. 2001, we see the year twice in the citation.
* Statement of need:
   * "such as 'Software Carpentry'," --> place the comma inside the quote.
   * "The target audience ... are researchers" --> *is*
* Learning objectives: "In addition to these" --> *the*
* Description of the module:
   * "The core of the course are" --> *is*
   * "to fit these functional form" --> *forms* (or *the* instead of *these*)
   *  "sum of square error objective function" --> *sum-of-square-error* (hyphens for compound adjective)
