### 2018-07-03 #16 @oliviaguest
Hi @labarba I'm ticking it on the proviso that this (https://github.com/arokem/scipy-optimize/issues/4) is all sorted, which I assume it is?
