### 2018-07-03 arokem/scipy-optimize#4 @labarba
As a matter of style, hyperlinked text should make sense when read, say, on a printed page, when the link is not active. So, instead of `methods see [this paper](...)`, one should write `methods, see the [paper by Smith]()`—or something such. In other words, avoid hyperlinked text like "this paper" or "here," because they don't make sense when read out.
