### 2018-07-03 6c98754
Fix links to read well on a page. Add references in reference section.
Addresses https://github.com/arokem/scipy-optimize/issues/4#issuecomment-402143360
