### 2018-07-03 arokem/scipy-optimize#4 @arokem
Thanks! I think this should be addressed in https://github.com/arokem/scipy-optimize/commit/6c987542a0a326e477b1715eff85c5f075cda22e
