### 2018-07-03 214ae5e
Fixes in paper.md
Addresses https://github.com/openjournals/jose-reviews/issues/16#issuecomment-401978516
