### 2018-07-04 578ca4f
Get rid of the year, in hopes of eliminating duplicated year.
