+++ paper.bib
@@ -11,7 +11,5 @@
 @Misc{scipy,
   author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001-},
   url = "http://www.scipy.org/",
-  note = {[Online; accessed <today>]}
 }
