+++ paper.bib
@@ -11,5 +11,6 @@
 @Misc{scipy,
   author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--},
   url = "http://www.scipy.org/",
 }
