### 2018-07-04 #16 @arokem
@labarba : do you understand why the year 2001 appears twice in the rendered pdf? It only appears once in [the bib file](https://github.com/arokem/scipy-optimize/blob/gh-pages/paper.bib#L11-L16)

And if I remove that, I get no year at all... (see the recent compilation attempt). Maybe whedon doesn't know how to properly handle the "misc" bib category?
