### 2018-07-04 #16 @labarba
@arfon, I wonder if you could help here. We're having a little problem with one of the references. See the comment by @arokem, [above](https://github.com/openjournals/jose-reviews/issues/16#issuecomment-402325346).
