### 2018-07-04 #16 @arfon
We had to this the other day with another bibtex file for an `online` entry
type. I’m afraid I don’t know what the fix is here (we didn’t figure it out
on the other submission)

On Tue, Jul 3, 2018 at 20:20, Lorena A. Barba <notifications@github.com>
wrote:

> @arfon <https://github.com/arfon>, I wonder if you could help here. We're
> having a little problem with one of the references. See the comment by
> @arokem <https://github.com/arokem>, above
> <https://github.com/openjournals/jose-reviews/issues/16#issuecomment-402325346>
> .
>
