### 2018-07-04 #16 @arokem
Yep, removing the dashes does remove the year duplication. Should I create a new archive for this change?
