### 2018-07-04 #16 @labarba
Can you just up the version of the archive on Zenodo? (keeps the DOI)
