### 2018-07-04 #16 @labarba
@arokem Note that Zenodo grabs the authors automatically from the GitHub repo. You need to manually change the author list there, since you get people listed that are not authors of this module.
