### 2018-07-04 #16 @labarba
@arokem Your paper is now published, yippee!!!
Please sign up as reviewer 😁

@oliviaguest, @ThomasA — Thank you again for reviewer. Do sign up (if you haven't) to our official reviewer list, for future requests, and help us advertise JOSE!
