### 2018-07-05 #16 @arokem
Thank you! Yippee!
On Wed, Jul 4, 2018 at 12:38 PM Lorena A. Barba <notifications@github.com>
wrote:

> @arokem <https://github.com/arokem> Your paper is now published, yippee!!!
> Please sign up as reviewer 😁
>
> <https://github.com/oliviaguest>
>
I believe I already did!

> @oliviaguest <https://github.com/oliviaguest>, @ThomasA
> <https://github.com/ThomasA> — Thank you again for reviewer. Do sign up
> (if you haven't) to our official reviewer list, for future requests, and
> help us advertise JOSE!
>
> Thanks to you both for the review!
